Initial commit for django-check-admin
django-check-admin is a Django app that adds a system check to verify
that all models have been registered with the Django admin site. This
check is useful if all models of a project should be registered by
default.

Use the Django management command `check`:

  $ python manage.py check

If a model of the project is not registered with the Django admin site,
an error will be displayed. For example:

  The model myapp.MyModel is not registered with an admin site.